Allow weights to be saved and loaded (#11)
* Add weights_file command line option and function to read weights (as floats) from file

* Move weighting code into weights.c

* Move main function out to main.c for if we ever want to add unit tests :)

* Add openMP option for mac

* Move argument parsing from run_plmc() to main()

* Add option to save weights to file

* Cleanup weights reading functions

* Use `brew --prefix` for openMP lib path

* Increase precision of saved weights

* Add regression test for weights and couplings

* Update README.md
